# docs/frontend-a/swagger.backend.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación realizada correctamente"
          },
          "400": {
            "description": "Solicitud inválida o reglas de negocio incumplidas"
          },
          "401": {
            "description": "Token JWT ausente, inválido o expirado"
          },
          "403": {
            "description": "No tienes permisos suficientes"
          },
          "404": {
            "description": "Proyecto no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar un proyecto administrativo",
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "operationId": "ProjectsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación realizada correctamente"
          },
          "400": {
            "description": "Solicitud inválida o reglas de negocio incumplidas"
          },
          "401": {
            "description": "Token JWT ausente, inválido o expirado"
          },
          "403": {
            "description": "No tienes permisos suficientes"
          },
          "404": {
            "description": "Proyecto no encontrado"
          },
          "409": {
            "description": "El slug ya está registrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Editar un proyecto administrativo",
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/api/v1/projects/{projectId}/milestones/{milestoneId}/contributions": {
      "post": {
        "operationId": "ContributionsController_recordMilestoneContributions",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "milestoneId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveContributionsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contribuciones registradas exitosamente"
          },
          "400": {
            "description": "Suma de porcentajes inválida o integrante no pertenece"
          },
          "403": {
            "description": "Solo el PO o Admin puede registrar contribuciones"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registrar contribuciones del equipo para un hito (PO)",
        "tags": [
          "Project contributions"
        ]
      },
      "get": {
        "operationId": "ContributionsController_getMilestoneContributions",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "milestoneId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar contribuciones de un hito",
        "tags": [
          "Project contributions"
        ]
      }
    },
...
    "/api/v1/projects/{projectId}/contributions": {
      "get": {
        "operationId": "ContributionsController_getProjectContributions",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resumen de contribuciones y desglose del equipo"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar reporte acumulado de contribuciones del proyecto",
        "tags": [
          "Project contributions"
        ]
      }
    }
  },
  "info": {
    "title": "TISNET API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sistema de inventario"
          },
          "slug": {
            "type": "string",
            "example": "sistema-de-inventario"
          },
          "shortDescription": {
            "type": "string",
            "example": "Control centralizado de productos y existencias."
          },
          "description": {
            "type": "string",
            "example": "Descripción pública completa del proyecto."
          },
          "problem": {
            "type": "object",
            "nullable": true
          },
          "solution": {
            "type": "object",
            "nullable": true
          },
          "objective": {
            "type": "object",
            "nullable": true
          },
          "features": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "categoryId": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "technologyIds": {
            "default": [],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "IN_DEVELOPMENT",
              "IN_REVIEW",
              "COMPLETED",
              "ARCHIVED"
            ],
            "default": "DRAFT"
          },
          "developmentDate": {
            "type": "object",
            "example": "2026-09-04",
            "nullable": true
          },
          "clientName": {
            "type": "object",
            "nullable": true
          },
          "demoUrl": {
            "type": "object",
            "example": "https://demo.example.com",
            "nullable": true
          },
          "externalUrl": {
            "type": "object",
            "example": "https://example.com",
            "nullable": true
          },
          "coverImageUrl": {
            "type": "object",
            "example": "https://cdn.example.com/cover.webp",
            "nullable": true
          },
          "isFeatured": {
            "type": "boolean",
            "default": false
          },
          "displayOrder": {
            "type": "number",
            "default": 0,
            "minimum": 0
          }
        }
      },
...
      "MemberContributionItemDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 10,
            "description": "ID del usuario integrante del equipo"
          },
          "percentage": {
            "type": "number",
            "example": 70,
            "description": "Porcentaje de contribución (1-100)"
          },
          "description": {
            "type": "string",
            "example": "Desarrollo de backend comercial y contratos",
            "description": "Descripción de las tareas realizadas"
          }
        },
        "required": [
          "userId",
          "percentage",
          "description"
        ]
      },
      "SaveContributionsDto": {
        "type": "object",
        "properties": {
          "contributions": {
            "description": "Lista de contribuciones por integrante",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberContributionItemDto"
            }
          }
        },
        "required": [
          "contributions"
        ]
      }
    }
  }
}